purviewblobtrigger/requirements.txt
azure-functions
azure-identity
requests
azure-storage-blob
azure-core
azure-servicebus
